fixed get_coord_avg_dict errors with multi-state pdbs
When multi-state pdb files are read into IMP, the pids of the copies of the protein will not be the same regardless of if they are identical structures. get_coord_avg_dict assumed that the pids representing each copy were the same, as would be the case if the structures were read individually than concatenated into a list.